Consolidate init of LAF colors into DockingUI
